Resolve the unknown namespace parameter.clj error
CIDER (and lein repl) was unable to load parameter.clj because it
depended on utils.clj, which in turn required the external opencv .so
library to be loaded.

It happens that the project.clj injections didn't work with lein repl
and thus with CIDER as well.

Adding a (load-opencv-libraries) free in the utils.clj code helped,
although not an ideal solution.